services/solr-names-updater/requirements/prod.txt
Flask
jsonschema
python-dotenv
sentry-sdk[flask]
aiohttp
asyncio-nats-client
asyncio-nats-streaming
Werkzeug
